# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "mem9 Server API",
    "version": "v1alpha2",
    "description": "OpenAPI specification for the mem9 Go server routes registered by server/internal/handler/handler.go."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.mem9.ai",
      "description": "Production API server"
    }
  ],
...
  "paths": {
...
    "/v1alpha1/mem9s": {
      "post": {
        "tags": [
          "Provisioning"
        ],
        "operationId": "provisionMem9s",
        "summary": "Provision a mem9 tenant",
        "description": "Creates a new tenant. The endpoint has no request body. Optional UTM query parameters are captured when present.",
        "parameters": [
          {
            "$ref": "#/components/parameters/UtmSource"
          },
          {
            "$ref": "#/components/parameters/UtmMedium"
          },
          {
            "$ref": "#/components/parameters/UtmCampaign"
          },
          {
            "$ref": "#/components/parameters/UtmContent"
          }
        ],
        "responses": {
          "201": {
            "description": "Tenant provisioned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1alpha2/status": {
      "get": {
        "tags": [
          "Status"
        ],
        "operationId": "getKeyStatus",
        "summary": "Get API key status",
        "description": "Validates X-API-Key against the control-plane state and returns whether the key is active or inactive.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "API key status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1alpha1/mem9s/{tenantID}/memories": {
      "post": {
        "tags": [
          "Memories v1alpha1"
        ],
        "operationId": "createTenantMemory",
        "summary": "Create or ingest memory data for a tenant",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantID"
          },
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateMemory"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/IngestOk"
          },
          "201": {
            "description": "Pinned memory created from explicit content.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryCreateResponse"
                }
              }
            }
          },
          "202": {
            "$ref": "#/components/responses/IngestAccepted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/RuntimeAccessBlocked"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/MemoryRouteRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      },
      "get": {
        "tags": [
          "Memories v1alpha1"
        ],
        "operationId": "listTenantMemories",
        "summary": "List or search tenant memories",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantID"
          },
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          },
          {
            "$ref": "#/components/parameters/MemoryQuery"
          },
          {
            "$ref": "#/components/parameters/MemorySearchMode"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Tags"
          },
          {
            "$ref": "#/components/parameters/Source"
          },
          {
            "$ref": "#/components/parameters/State"
          },
          {
            "$ref": "#/components/parameters/MemoryType"
          },
          {
            "$ref": "#/components/parameters/AgentID"
          },
          {
            "$ref": "#/components/parameters/SessionID"
          },
          {
            "$ref": "#/components/parameters/AppID"
          },
          {
            "$ref": "#/components/parameters/ScanAll"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MemoryList"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/RuntimeAccessBlocked"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/MemoryRouteRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1alpha1/mem9s/{tenantID}/memories/{id}": {
      "get": {
        "tags": [
          "Memories v1alpha1"
        ],
        "operationId": "getTenantMemory",
        "summary": "Get a tenant memory by ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantID"
          },
          {
            "$ref": "#/components/parameters/MemoryID"
          },
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Memory"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "put": {
        "tags": [
          "Memories v1alpha1"
        ],
        "operationId": "updateTenantMemory",
        "summary": "Update a tenant memory",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantID"
          },
          {
            "$ref": "#/components/parameters/MemoryID"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdateMemory"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MemoryWithETag"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/RuntimeAccessBlocked"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/MemoryRouteRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "Memories v1alpha1"
        ],
        "operationId": "deleteTenantMemory",
        "summary": "Delete a tenant memory",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantID"
          },
          {
            "$ref": "#/components/parameters/MemoryID"
          },
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          }
        ],
        "responses": {
          "204": {
            "description": "Memory deleted."
          },
          "402": {
            "$ref": "#/components/responses/RuntimeAccessBlocked"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/MemoryRouteRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1alpha1/mem9s/{tenantID}/imports": {
      "post": {
        "tags": [
          "Imports v1alpha1"
        ],
        "operationId": "createTenantImport",
        "summary": "Create an async tenant import task",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantID"
          },
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateImport"
        },
        "responses": {
          "202": {
            "$ref": "#/components/responses/TaskCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "get": {
        "tags": [
          "Imports v1alpha1"
        ],
        "operationId": "listTenantImports",
        "summary": "List tenant import tasks",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantID"
          },
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TaskList"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1alpha1/mem9s/{tenantID}/imports/{id}": {
      "get": {
        "tags": [
          "Imports v1alpha1"
        ],
        "operationId": "getTenantImport",
        "summary": "Get a tenant import task",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantID"
          },
          {
            "$ref": "#/components/parameters/TaskID"
          },
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TaskDetail"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1alpha1/mem9s/{tenantID}/session-messages": {
      "get": {
        "tags": [
          "Sessions v1alpha1"
        ],
        "operationId": "listTenantSessionMessages",
        "summary": "List raw session messages for one or more sessions",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantID"
          },
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          },
          {
            "$ref": "#/components/parameters/SessionIDRepeated"
          },
          {
            "$ref": "#/components/parameters/AppID"
          },
          {
            "$ref": "#/components/parameters/LimitPerSession"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SessionMessages"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1alpha2/mem9s/runtime-state": {
      "get": {
        "tags": [
          "Runtime State v1alpha2"
        ],
        "operationId": "getRuntimeState",
        "summary": "Get plugin-visible runtime quota state",
        "description": "Returns runtime-state facts for agent plugins. This route uses local API-key status resolution and does not depend on tenant data DB resolution. Hosted deployments read provider state through the configured runtime usage provider. Provider-disabled self-host deployments return a local fallback response with the same public shape.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/RuntimeState"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1alpha2/mem9s/memories": {
      "post": {
        "tags": [
          "Memories v1alpha2"
        ],
        "operationId": "createMemory",
        "summary": "Create or ingest memory data",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateMemory"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/IngestOk"
          },
          "201": {
            "description": "Pinned memory created from explicit content.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryCreateResponse"
                }
              }
            }
          },
          "202": {
            "$ref": "#/components/responses/IngestAccepted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/RuntimeAccessBlocked"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/MemoryRouteRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      },
      "get": {
        "tags": [
          "Memories v1alpha2"
        ],
        "operationId": "listMemories",
        "summary": "List or search memories",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          },
          {
            "$ref": "#/components/parameters/MemoryQuery"
          },
          {
            "$ref": "#/components/parameters/MemorySearchMode"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Tags"
          },
          {
            "$ref": "#/components/parameters/Source"
          },
          {
            "$ref": "#/components/parameters/State"
          },
          {
            "$ref": "#/components/parameters/MemoryType"
          },
          {
            "$ref": "#/components/parameters/AgentID"
          },
          {
            "$ref": "#/components/parameters/SessionID"
          },
          {
            "$ref": "#/components/parameters/AppID"
          },
          {
            "$ref": "#/components/parameters/SortBy"
          },
          {
            "$ref": "#/components/parameters/SortDir"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MemoryList"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/RuntimeAccessBlocked"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/MemoryRouteRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/v1alpha2/mem9s/imports/{id}": {
      "get": {
        "tags": [
          "Imports v1alpha2"
        ],
        "operationId": "getImport",
        "summary": "Get an import task",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TaskID"
          },
          {
            "$ref": "#/components/parameters/XMnemoAgentId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/TaskDetail"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key used by v1alpha2 routes."
      }
    },
    "parameters": {
...
      "UtmSource": {
        "name": "utm_source",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "UtmMedium": {
        "name": "utm_medium",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "UtmCampaign": {
        "name": "utm_campaign",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "UtmContent": {
        "name": "utm_content",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request. For v1alpha2 API-key routes this also includes missing, invalid, or inactive X-API-Key values returned by the authentication middleware.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Version, duplicate-key, or resource conflict.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ProvisionResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Provisioned tenant ID."
          }
        }
      },
...
    }
  }
}